finaly feature user address

+++ backEnd/app/Http/Controllers/UserAddressController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\UserAddressRequest;
 use App\Models\User;
+use App\Models\UserAddress;
 use App\Repositories\UserAddressRepositoryInterface;
 use Exception;
 use Illuminate\Http\Request;
@@ -11,14 +12,17 @@ use Illuminate\Http\Request;
 class UserAddressController extends Controller
 {
     private $userAddressRepo;
-    public function __construct(UserAddressRepositoryInterface $userAddressRepo)
+    private $user;
+    public function __construct(UserAddressRepositoryInterface $userAddressRepo, Request $request)
     {
         $this->userAddressRepo = $userAddressRepo;
+        $this->user = $request->user();
     }
     public function index()
     {
+        $id = $this->user->id;
         try {
-            $result['data'] = $this->userAddressRepo->getAll();
+            $result['data'] = $this->userAddressRepo->getByUserId($id);
         } catch (Exception $e) {
             $result = [
                 'status'    => 500,
@@ -29,8 +33,9 @@ class UserAddressController extends Controller
     }
     public function store(UserAddressRequest $requst)
     {
+        $userId = $this->user->id;
         try {
-            $result['data'] = $this->userAddressRepo->create($requst->all());
+            $result['data'] = $this->userAddressRepo->create($userId, $requst->all());
         } catch (Exception $e) {
             $result = [
                 'status'    => 500,
@@ -41,6 +46,7 @@ class UserAddressController extends Controller
     }
     public function show($id)
     {
+        $this->authorize('userAddressPolicy', $this->userAddressRepo->getById($id));
         try {
             $result['data'] = $this->userAddressRepo->getById($id);
         } catch (Exception $e) {
@@ -53,6 +59,7 @@ class UserAddressController extends Controller
     }
     public function update(UserAddressRequest $requst, $id)
     {
+        $this->authorize('userAddressPolicy', $this->userAddressRepo->getById($id));
         try {
             $result['data'] = $this->userAddressRepo->update($id, $requst->all());
         } catch (Exception $e) {
@@ -65,6 +72,7 @@ class UserAddressController extends Controller
     }
     public function destroy($id)
     {
+        $this->authorize('userAddressPolicy', $this->userAddressRepo->getById($id));
         try {
             $result['data'] = $this->userAddressRepo->delete($id);
         } catch (Exception $e) {
